+++ common/src/error_handler.rs
@@ -1,0 +1,22 @@
+use super::*;
+
+pub async fn on_error(error: poise::FrameworkError<'_, Data, Error>) {
+    // This is our custom error handler
+    // They are many errors that can occur, so we only handle the ones we want to customize
+    // and forward the rest to the default handler
+    // これはカスタム エラー ハンドラーです
+    // 多くのエラーが発生する可能性があるため、カスタマイズしたいエラーのみを処理します
+    // そして残りをデフォルトのハンドラーに転送します
+    match error {
+        poise::FrameworkError::Setup { error, .. } => panic!("Failed to start bot: {:?}", error),
+        poise::FrameworkError::Command { error, ctx, .. } => {
+            info!("Error in command `{}`: {:?}", ctx.command().name, error,);
+            ctx.say(error.to_string()).await.ok();
+        }
+        error => {
+            if let Err(e) = poise::builtins::on_error(error).await {
+                info!("Error while handling error: {}", e)
+            }
+        }
+    }
+}
